Use maxHeight from Popover to make popover not overflow
Ugly, but makes the Popover usable. Wondering if these components would
be more usable inlined in the Sidebar as we need to maintain enough
height to support showing list of options from TokenInput as well as the
button to add the filter. Current solution is to just force the button
to appear above the list.