Update semantic migrations for mysql compatibility
mysql updates as it goes, not in a row level transaction. So

```sql
update a = null,
       b = (case a ...)
where a is not null
```

mysql will use the new null value of a rather than the pre-udpate
value of a whereas pg and h2 will do row level transaction so you are
considering the pre-updated value of a.

Solution is to update these migrations to leave semantic type alone
and then one final set semantic_type = null where coercion_strategy is
not null. This is sensical since these two properties used to both be
in semantic_type, so if we extracted a coercion it could not have had
a meaningful semantic_type before.