Check columns for datetime behavior based on special_type (#15319)
master has a semantic_type that this should be updated to when
merged. Or even better, if columns have an effective type we can just
use that. I'm not sure how inference goes so it may just have a
base_type that is the datetime type if its just inferred, or if the
column is recognized as a field maybe it has both base and effective
types