Move operator information into shared cljc file
simplify the definitions of parameters on the frontend. It had a flat
list and then would try to group things according to the name and
dynamically add them in places. I think this is easier. Now there's
number/string/date/id type parameters and the frontend can just
declare which ones it cares about. It was quite complicated for some
reason before.

Added a `trs` macro on the frontend so i could move some functions
into the shared bits.

Technically, now the frontend could generate the identical mbql the
backend will run.

TODO: move those date style things into the backend. There's a ticket
to clean up the old style parameters and at that point this all gets
moved to the backend.