Check can connect to db and serialize syncs
rather than allowing us to create threads willy nilly, serialize them
and reuse them in a fixed thread pool with a single thread. This helps
if someone "multiple clicks" and launches a few table syncs. It's not
appropriate to use the existing deduplicating stuff in the sync which
prevents later tasks from happening while the current sync step
happens since we want to sync with the most recent changes. So just
serialize them to go one at a time rather than off to the races.