Restore table visibility (#15916)
* Move unhidden table sync off-thread

* Test bulk hide/unhide syncing

* Check can connect to db and serialize syncs

rather than allowing us to create threads willy nilly, serialize them
and reuse them in a fixed thread pool with a single thread. This helps
if someone "multiple clicks" and launches a few table syncs. It's not
appropriate to use the existing deduplicating stuff in the sync which
prevents later tasks from happening while the current sync step
happens since we want to sync with the most recent changes. So just
serialize them to go one at a time rather than off to the races.

* Clean up ns, make private, and type hints

* Block waiting on table sync to finish

* Use driver.util can connect helper

uses the timeout of 5seconds and catches errors

* disable ToggleHiddenButton during request

Co-authored-by: Dalton Johnson <daltojohnso@users.noreply.github.com>