Fall back to base_type when calling parse-value-for-field-type (#15905)
I don't exactly follow what's going on in #15901. But it seems that
the field information can be kept in the db

> ""field_ref"":[""field"",""count(*)"",{""base-type"":""type/BigInteger""}]

And so we need to handle this base type here. Note that in the
function that calls `update-filter-for-field-type` it checks that the
field has a base_type.

I'm also a bit confused by the base_type vs base-type in the above
snippet. But again the calling code has

```clojure
(and (= param-type :dimension)
     (get-in value [:field :base_type]))
(update-filter-for-field-type value)
```

So it seems the field has a base_type rather than a base-type.