Parse number filter inputs
The frontend feels a bit off on this. It's sending `"string/="`
instead of `"number/="` but the backend code knows that number filters
can ignore it. It's also sending strings from the text inputs when it
should be sending number types.

We've had a convention that the types come across as they should be:
numbers come in as numbers and strings as strings. This seems to break
that convention.