Don't case in the outer select when inner select will
Expressions create subselects

```clojure
(-> {:database 1,
     :query {:source-table 2
             :expressions {:test ["*" 1 1]}
             :fields [[:field 2 nil]
                      [:expression "test"]]}
     :type "query"
     :parameters []}
    qp/process-query
    (update-in [:data :rows] count)
    pprint)
```

```sql
SELECT "source"."QUANTITY" AS "QUANTITY",
       "source"."test" AS "test"
FROM (SELECT timestampadd('second', "PUBLIC"."ORDERS"."QUANTITY", timestamp '1970-01-01T00:00:00Z')
     AS "QUANTITY",
     (1 * 1) AS "test"
     FROM "PUBLIC"."ORDERS") "source"
LIMIT 1048575
```

The bug was that source quantity was marked as a coerced integer to
timestamp, and it coerced in both the inner and outer select.

There's of course a few ways to solve this:
- the easiest way is when fetching the field information, throw away
the coercion strategy. Didn't want to do this as less information gets
confusing. Preferred to add information about _why_ it should be
ignored rather than just suppressing information.
- (solution here) add a marker to just not coerce on the outer
select.

The metadata is a bit strange to me. I'm honestly not sure what it
should be. What is the base type, effective type, and coercion
strategies of the things returned? At the moment they are:

```clojure
{:semantic_type :type/Quantity,
 :coercion_strategy :Coercion/UNIXSeconds->DateTime,
 :name "QUANTITY",
 :field_ref [:field 2 nil],
 :effective_type :type/Instant
 :base_type :type/DateTime}
```

Should this still have coercion information and a base type of
integer (ie, promote the inner field type's information to the
toplevel) or is this properly the type information now?