Fix standard filter detection
When attempting to convert a filter, check first if the filter is a
comparison against a constant. Otherwise, preserve the filter is in its
custom expression (MBQL) form.

Each in the following list can be converted into a standard filter
(since the simple UI for the standard filter only allows literal
comparison):

  [Total] < 300
  between([Discount], 5, 15)
  [Name] = "Joe"

but none of these should be converted

  [Discount] <= [Tax]
  between([Rating], 0, [MaxRating])
  [Reviewer] != UPPER([Name])