Show revert button in timeline and make it reload the card
the reloadCard action previously retrieved an "originalCard" from state.
I'm not totally clear _why_, but it seems like the correct course of
action after triggering a revert is to re-GET the card so that it
reflects the revert. It may be a revert from a long time ago and that
would not be stored in client state.