Parse number filter inputs (#15992)
* Parse number filter inputs

The frontend feels a bit off on this. It's sending `"string/="`
instead of `"number/="` but the backend code knows that number filters
can ignore it. It's also sending strings from the text inputs when it
should be sending number types.

We've had a convention that the types come across as they should be:
numbers come in as numbers and strings as strings. This seems to break
that convention.

* Unskip the "number" part of the repro for #15981

* use normalized parameter values when diffing for dirtiness

* update tag parameter parsing fns to handle number --> number/= types

* tweak & unskip cypress repro test

* wrap fn in parens to invoke it?

Co-authored-by: Nemanja <31325167+nemanjaglumac@users.noreply.github.com>
Co-authored-by: Dalton Johnson <daltojohnso@users.noreply.github.com>