Fix standard filter detection (#16174)
When attempting to convert a filter, check first if the filter is a
comparison against a constant. Otherwise, preserve the filter is in its
custom expression (MBQL) form.

Each in the following list can be converted into a standard filter
(since the simple UI for the standard filter only allows literal
comparison):

  [Total] < 300
  between([Discount], 5, 15)
  [Name] = "Joe"

but none of these should be converted

  [Discount] <= [Tax]
  between([Rating], 0, [MaxRating])
  [Reviewer] != UPPER([Name])

* Allow standard filter to be invalid

This is necessary to allow an incomplete filter, i.e. the user hasn't
filled Y in "X < Y" form yet (hence Y is undefined). The filter is not
valid but it is still considered in the standard form.

* Check whether string comparison can be considered a standard filter

* Idiomatic with array destructuring