Add the last-edit-info to the child query
just for cards right now, need to extend to dashboards and then make
sure we clean it up into the proper map.

Have a set of all columns that are required and ensure those are in
each query. Nulls need to be cast to the correct type if they aren't a
text column (at least on postgres, need to investigate h2 and mysql).