Simple question summary sidebar: highlight the selected binning (#16665)
This mimics the situation in the notebook (custom question). A minor
difference, hence the omission of the previous implementation, is that
notebook breakout selection only allows one table at a time (the other
tables are "folded" in the accordion list component). In the case of a
simple question, every breakout selection happens in a sidebar, a long
list of all tables and all dimensions. Thus, it needs to support
multiple selection at once, which this commit finally implements.