Don't call completion arity from step arity
the `(rf acc)` call violated the terms of transduction. If it does
need to be there, perhaps it could be used with reduced

wary of #12207 (which fixes #12165) and seeing if that crops up again.

- i had introduced :metrics/ignore-execution-time but this was because
i thought that we were running a simple query (limit 1) to get col
metadata. Now that i know we were just being a bit careless with `(rf
acc)` this extra information is no longer needed, and we can just
react to cached data.