Cache results dont affect average execution time (#16720)
* Only save successful-query-execution if it wasnt a cached result

the only thing this function does is update the rolling average query
execution time

* Failing tests for cache execution stats

These tests fail, but they shouldn't.

The save-execution information should not be called twice, and the
average execution duration should remain the same. What's happening is
that it is getting called twice just in the cache call, and only one
of those has the new key `:metrics/ignore-execution-time`. This
remains a mystery to me

* Don't call completion arity from step arity

the `(rf acc)` call violated the terms of transduction. If it does
need to be there, perhaps it could be used with reduced

wary of #12207 (which fixes #12165) and seeing if that crops up again.

- i had introduced :metrics/ignore-execution-time but this was because
i thought that we were running a simple query (limit 1) to get col
metadata. Now that i know we were just being a bit careless with `(rf
acc)` this extra information is no longer needed, and we can just
react to cached data.

* hoist closing paren