Add seen optimization of nodes
We reach nodes in the shortest path possible, so keep a list of nodes
visited and don't revisit them: we got to them as fast as possible,
and once there, there's no new nodes that we can see from that vantage
point.

This allows a serious speedup and should hopefully fully resolve the
cpu issues seen by some.