Rename CI group `filters` -> `native-filters` (#17066)
This group already takes close to 10 minutes to finish.

We need to have an additional extended coverage for dashboard filters and for query builder filters.
It doesn't make sense to keep them all together under the `filters` CI group because that group on its own could take more than half an hour to finish.