Remove @babel/standalone and dependent code
We're using @babel/standalone to support the writing of JSX in our
internal-only ScratchApp editor. Unfortunately, this dependency is
large--1.5mb not gzipped--and our current build does not intelligently
split bundles, meaning our users our burdened by the size of this
dependency whenever they must redownload the Metabase JS bundle.

I'm removing it. We should still be able to write JS in the ScratchApp,
but JSX no longer.