Indicate that variable names must be alphanumeric
If you use a non-alphanumeric variable name like `variable_name` in the docs, the variable does not appear in the side bar which is quite confusing. There is a comment in the code that indicates only alphanumeric values are accepted, so I added this line in the docs to reflect that. https://github.com/metabase/metabase/blob/master/frontend/src/metabase-lib/lib/queries/NativeQuery.js#L378

It also says another form is allowed, but I'm not sure that it should be in the documentation.