Include depstar
osx has an error when building the jar:
- the osx filesystem is case insensitive. So cannot have a folder and
a file with the same name except for casing
- some jars had a file `LICENSE` and another `license/LICENSE`. But
the temp dir that tools.build uses to hold all of the contents cannot
hold these both on osx, even though linux and the jars can.

Depstar doesn't copy the contents of jars onto the filesystem, just
consumes from the jar. So it always exists in a case insensitive
location.